// frontend/src/app/services/models/page-response-dto-tool-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ToolResponseDto } from '../models/tool-response-dto';
export interface PageResponseDtoToolResponseDto {
  content?: Array<ToolResponseDto>;
  first?: boolean;
  last?: boolean;
  page?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}